Tweak cache strategy

+++ clean_disk_cache.sh
@@ -19,10 +19,18 @@ set -eu
 
 BAZEL_DISK_CACHE_PATH=$1
 
-# As a courtesy, compute and print some approximate stats.
-total_file_count=$(find "$BAZEL_DISK_CACHE_PATH" -type f | wc -l)
-stale_file_count=$(find "$BAZEL_DISK_CACHE_PATH" -type f -atime +3 | wc -l)
-echo "Removing $stale_file_count files out of $total_file_count total."
+SIZE =$(du -sk "$BAZEL_DISK_CACHE_PATH" | cut -f1)
 
-# Just re-running the find is simpler than managing any state.
-find "$BAZEL_DISK_CACHE_PATH" -type f -atime +3 -delete
+echo "Bazel cache dir size = $SIZE KB"
+
+# 1.5GB = 1572864KB
+if [[ $SIZE -gt 1572864 ]]; then
+    echo "cache dir over size, try pruning"
+    # As a courtesy, compute and print some approximate stats.
+    total_file_count=$(find "$BAZEL_DISK_CACHE_PATH" -type f | wc -l)
+    stale_file_count=$(find "$BAZEL_DISK_CACHE_PATH" -type f -atime +3 | wc -l)
+    echo "Removing $stale_file_count files out of $total_file_count total."
+
+    # Just re-running the find is simpler than managing any state.
+    find "$BAZEL_DISK_CACHE_PATH" -type f -atime +3 -delete
+fi
